requirements.txt
six
nibabel
numpy
scipy
medpy
tensorflow-gpu
configparser
scikit-image
